shipment_service/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==3.2.2
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
motor==3.7.0
passlib==1.7.4
pillow==11.1.0
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
python-barcode==0.15.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
